requirements.txt
# Основные зависимости для Telegram бота
python-telegram-bot==21.6
httpx==0.27.0

# База данных и кеширование
redis==5.0.1

# Интеграции для векторной системы
chromadb>=0.4.0
openai==1.12.0

# Утилиты
python-dotenv==1.0.0
pydantic==2.5.3
structlog==23.2.0

# Тестирование
pytest==7.4.3
pytest-asyncio==0.21.1
